Upload 1.18.2 Fabric mod project
This is the Discord Relay I used on my 1.18 community Minecraft server.

I previously worked on it locally and never bothered to publish it to GitHub as I was still new to using IntelliJ and Java programming for Minecraft mods so I just did not want to bother with extra complication through version control. However, as I am planning to use this on my 1.19 server and am more experienced with Minecraft and Fabric development, I have decided to finally publish this project.

The previous Paper plugin is now archived on a seperate branch, and this Fabric branch will be a fresh start.