New API class, rewrite config & utility method, etc.
Made a better way of creating and loading configuration files.

Added a new directory to hold everything Discord related (API, Gateway, etc.). Currently only the API class is implemented, which everything else needs to be changed over to use. The Discord API base URL and version needs to be added to the configuration file too. Maybe webhook configuration should just be ID and token too, and use the base URL provided in the aforementioned property.